requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
twilio
pymongo
requests
openai
pydantic
pillow
